+++ cvts/tasks/_regiondensity.py
@@ -4,7 +4,6 @@ import pickle
 from multiprocessing import Pool
 from functools import partial as _partial
 from datetime import timezone, timedelta
-from math import floor, ceil
 import logging
 import numpy as np
 from tqdm import tqdm
@@ -13,6 +12,7 @@ from .. import (
     read_shapefile,
     points_to_polys,
     distance)
+from .._grid import Grid
 from ..settings import (
     OUT_PATH,
     STOP_PATH,
@@ -40,71 +40,10 @@ MAGIC_DISTANCE = 50
 #: Timezone for Vietnam.
 TZ             = timezone(timedelta(hours=7), 'ITC')
 
-#: Minimum latitude of the raster 'covering' Vietnam.
-MINLAT   =   7.8584
-
-#: Approximate maximum latitude of the raster 'covering' Vietnam.
-MAXLAT   =  23.8882
-
-#: Minimum latitude of the raster 'covering' Vietnam.
-MINLON   = 101.9988
-
-#: Approximate maximum longitude of the raster 'covering' Vietnam.
-MAXLON   = 109.3325
-
-#: Cellsize of the raster 'covering' Vietnam.
-CELLSIZE =   0.1
-
-#: NA value to use for the raster 'covering' Vietnam.
-NA_VALUE = -9999
-
 POINTS_GEOM_IDS_FILE_POSTFIX        = 'points_geom_ids.pkl'
 POINTS_LON_LAT_FILE_POSTFIX         = 'points_lon_lat.pkl'
 POINTS_GEOM_COUNTS_FILE_POSTFIX     = 'points_geom_counts.pkl'
 POINTS_GEOM_COUNTS_CSV_FILE_POSTFIX = 'points_geom_counts.csv'
-
-
-
-class _Grid:
-    """Raster used for accumulating stop points."""
-
-    def __init__(
-            self,
-            minlat   = MINLAT,
-            minlon   = MINLON,
-            maxlat   = MAXLAT,
-            maxlon   = MAXLON,
-            cellsize = CELLSIZE,
-            na_value = NA_VALUE):
-        self.minlat = minlat
-        self.minlon = minlon
-        self.cellsize = cellsize
-        self.na_value = na_value
-        self.ncol = int(ceil((maxlon - self.minlon) / self.cellsize))
-        self.nrow = int(ceil((maxlat - self.minlat) / self.cellsize))
-        self.cells = np.zeros((self.nrow, self.ncol), int)
-
-    def increment(self, lon, lat):
-        """Increment the count in cell containing the point (*lon*, *lat*)."""
-        try:
-            row = self.nrow - int(floor((lon - self.minlon) / self.cellsize)) - 1
-            col =             int(floor((lat - self.minlat) / self.cellsize))
-            if 0 <= col < self.ncol and 0 <= row < self.nrow:
-                self.cells[row, col] += 1
-        except ValueError as e:
-            logger.warning('value error: {} at ({:.2f}, {:.2f})'.format(e, lat, lon))
-        # TODO: warning here?
-
-    def save(self, fn):
-        """Save as ASCII grid."""
-        with open(fn, 'w') as f:
-            f.write('ncols        {}\n'.format(self.ncol))
-            f.write('nrows        {}\n'.format(self.nrow))
-            f.write('xllcorner    {}\n'.format(self.minlon))
-            f.write('yllcorner    {}\n'.format(self.minlat))
-            f.write('cellsize     {}\n'.format(self.cellsize))
-            f.write('NODATA_value {}\n'.format(self.na_value))
-            f.write(' '.join([str(i) for i in self.cells.flatten()]))
 
 
 
@@ -387,7 +326,7 @@ class RasterCounts(luigi.Task):
             stop_points = pickle.load(inf)
 
         # construct and save the grid counts
-        grid = _Grid()
+        grid = Grid()
         for p in stop_points:
             grid.increment(*p)
         grid.save(self.output().fn)
